UIBaseクラスを追加

+++ FrameWork/FrameWork/Fighting/Main/SceneManager/SceneBase/GameScene/ObjectManager/ObjectBase/GameObjectBase/Player/HandBase/LeftHand/LeftHand.cpp
@@ -65,7 +65,7 @@ void LeftHand::Update()
 		}
 		else if (m_Pos.x < ClientRect.right / 2 - 300)
 		{
-			m_Pos.x = ClientRect.right / 2 - 300;
+			m_Pos.x = static_cast<float>(ClientRect.right / 2 - 300);
 		}
 	}
 }
